scripts/requirements.txt
attrs==25.4.0
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
fiona==1.9.6
folium==0.20.0
geopandas==0.14.4
h3==3.7.6
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
networkx==3.2.1
numpy==1.26.3
osmnx==1.9.2
packaging==25.0
pandas==2.1.4
pyarrow==15.0.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
Rtree==1.1.0
scipy==1.11.4
shapely==2.0.2
six==1.17.0
tqdm==4.66.4
tzdata==2025.2
urllib3==2.5.0
xyzservices==2025.11.0
